+++ mod/twitch/src/main/java/tv/twitch/android/models/channel/ChannelInfo.java
@@ -1,0 +1,5 @@
+package tv.twitch.android.models.channel;
+
+public interface ChannelInfo {
+    int getId();
+}