+++ mod/twitch/src/main/java/tv/twitch/android/shared/chat/observables/ChatConnectionController.java
@@ -1,0 +1,6 @@
+package tv.twitch.android.shared.chat.observables;
+
+public class ChatConnectionController {
+    public int viewerId;
+    public void BTTVdisconnect(int broadcasterId) {}
+}