+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# master
+# v0.19.0-beta
 - add anon chat
 - add some bttv-android settings to stream settings
 - fix bug related to Watch Parties